[iOS] refactor: 그룹 정보 데이터 소스 객체 분리
